+++ libs/wire-api/src/Wire/API/Team/Feature.hs
@@ -29,7 +29,7 @@ module Wire.API.Team.Feature
     KnownTeamFeatureName (..),
     TeamFeatureStatusNoConfig (..),
     TeamFeatureStatusWithConfig (..),
-    deprecatedFeatureName,
+    HasDeprecatedFeatureName (..),
     defaultAppLockStatus,
     defaultClassifiedDomains,
 
@@ -154,11 +154,17 @@ instance ToByteString TeamFeatureName where
   builder TeamFeatureAppLock = "appLock"
   builder TeamFeatureClassifiedDomains = "classifiedDomains"
 
-deprecatedFeatureName :: TeamFeatureName -> Maybe ByteString
-deprecatedFeatureName TeamFeatureSearchVisibility = Just "search-visibility"
-deprecatedFeatureName TeamFeatureValidateSAMLEmails = Just "validate-saml-emails"
-deprecatedFeatureName TeamFeatureDigitalSignatures = Just "digital-signatures"
-deprecatedFeatureName _ = Nothing
+class HasDeprecatedFeatureName (a :: TeamFeatureName) where
+  type DeprecatedFeatureName a :: Symbol
+
+instance HasDeprecatedFeatureName 'TeamFeatureSearchVisibility where
+  type DeprecatedFeatureName 'TeamFeatureSearchVisibility = "search-visibility"
+
+instance HasDeprecatedFeatureName 'TeamFeatureValidateSAMLEmails where
+  type DeprecatedFeatureName 'TeamFeatureValidateSAMLEmails = "validate-saml-emails"
+
+instance HasDeprecatedFeatureName 'TeamFeatureDigitalSignatures where
+  type DeprecatedFeatureName 'TeamFeatureDigitalSignatures = "digital-signatures"
 
 typeTeamFeatureName :: Doc.DataType
 typeTeamFeatureName = Doc.string . Doc.enum $ cs . toByteString' <$> [(minBound :: TeamFeatureName) ..]